Converted Cpu, Joypad and Interrupts to pure C
