+++ src/Cpu/Operations.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <cstdint>
-#include "Cpu/Cpu.hpp"
+#include "Cpu/Cpu.h"
 
 struct Instruction;
