+++ src/Gameboy/Gameboy.cpp
@@ -3,6 +3,7 @@
 #include "Gameboy.hpp"
 #include "Cpu/Opcodes.hpp"
 #include "Cpu/Interpreter.hpp"
+#include "Cpu/Interrupts.h"
 
 GameBoy::GameBoy(MemoryBankController* mc)
     : memory_bank_controller(mc) 
@@ -12,7 +13,6 @@ GameBoy::GameBoy(MemoryBankController* mc)
         joypad.direction_keys = 0b1111;
         joypad.button_keys = 0b1111;
         cpu.memory_controller = memory_bank_controller;
-        interrupt_handler.memory_bank_controller = memory_bank_controller;
 
         set_initial_state(&cpu);
     }
@@ -21,7 +21,6 @@ void GameBoy::run() {
 
     const auto opcode = read(memory_bank_controller, cpu.program_counter);
 
-
     auto instruction = fetch(static_cast<Opcode>(opcode));
 
     auto cycles = 4;
@@ -29,14 +28,14 @@ void GameBoy::run() {
         auto operand = instruction.read_operand(cpu);
         cycles = instruction.execute(cpu, operand);
     } else {
-        cpu.is_halted = interrupt_handler.should_exit_halt();
+        cpu.is_halted = should_exit_halt(memory_bank_controller);
     }
 
     timer.increment(cycles);
     ppu.step(cycles);
         
     if(cpu.interrupts_enabled) {
-        const auto interrupt_cycles = interrupt_handler.interrupts(cpu);
+        const auto interrupt_cycles = handle_interrupts(&cpu);
         timer.increment(interrupt_cycles);
     }
