+++ src/Gameboy/Gameboy.hpp
@@ -1,13 +1,10 @@
 #pragma once
 
 #include "Memory/Memory_Controller.h"
-
-#include "Cpu/Cpu.hpp"
+#include "Cpu/Cpu.h"
+#include "Input/Joypad.h"
 #include "Cpu/Timer.hpp"
-#include "Cpu/Interrupt_Handler.hpp"
 #include "Graphics/PPU.hpp"
-#include "Input/Joypad.hpp"
-
 
 struct GameBoy {
 
@@ -17,7 +14,6 @@ struct GameBoy {
     MemoryBankController* memory_bank_controller;
     Cpu cpu;
     Timer timer;
-    Interrupt_Handler interrupt_handler;
     PPU ppu;
     Joypad joypad;
 };
