+++ src/Graphics/PPU.cpp
@@ -1,14 +1,14 @@
 #include "Graphics/PPU.hpp"
 #include "Utilities/BitOperations.h"
 #include "Memory/Memory_Controller.h"
+#include "Cpu/Interrupts.h"
+
 
 PPU::PPU(MemoryBankController* memory_controller)
     : m_memory_controller(memory_controller)
     , m_lcd_control(m_memory_controller)
     , m_lcd_status(m_memory_controller, m_lcd_control)
-{
-    m_interrupt_handler.memory_bank_controller = m_memory_controller;
-}
+{}
 
 void PPU::step(int cycles) {
